Replace fiberNormsQa with takahashiFiberNormsQa
