+++ bin.src/fiberNormsQa.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from pfs.drp.qa.takahashiFiberNormsQa import main
+from pfs.drp.qa.fiberNormsQa import main
 
 if __name__ == "__main__":
     main()
